feat:record 수정확인
