+++ kosta_finalpjt/src/main/java/com/example/demo/config/oracleDBConfig.java
@@ -62,6 +62,7 @@ public class oracleDBConfig {
     return DataSourceBuilder.create()
         .driverClassName("oracle.jdbc.driver.OracleDriver")
         .url("jdbc:oracle:thin:@192.168.0.36:1521/xe")
+//        .url("jdbc:oracle:thin:@localhost:1521/xe")
         .username("hr")
         .password("hr")
         .build();
