+++ kosta_finalpjt/src/main/java/com/example/demo/oracledb/workinoutrecords/WorkInOutRecordService.java
@@ -48,6 +48,7 @@ public class WorkInOutRecordService {
 		
 	//부서 직원 통계기록
 	public ArrayList<ChartDeptMember> chartMonthandDept(int month,int year,int dept){
+		System.out.println("부서 직원 통계");
 		List<Object[]> deptlist = dao.chartDept(month, year, dept);
 		ArrayList<ChartDeptMember> list = new ArrayList<ChartDeptMember>();
 		for(Object[] a: deptlist) {
