+++ kosta_finalpjt/src/main/resources/static/js/record/record_common.js
