+++ kosta_finalpjt/src/main/resources/templates/record/dept.html
@@ -98,6 +98,6 @@
 
 </div>
 </body>
-<script th:src="@{/js/record_common.js}"></script>
-<script th:src="@{/js/record_dept.js}"></script>
+<script th:src="@{/js/record/record_common.js}"></script>
+<script th:src="@{/js/record/record_dept.js}"></script>
 </html>
