+++ kosta_finalpjt/src/main/resources/templates/record/my.html
@@ -113,6 +113,6 @@
   </div>
 
 </body>
-<script th:src="@{/js/record_common.js}"></script>
-<script th:src="@{/js/record_emp.js}"></script>
+<script th:src="@{/js/record/record_common.js}"></script>
+<script th:src="@{/js/record/record_emp.js}"></script>
 </html>
